requirements/dev.txt
-r base.txt

black==19.3b0
ipdb==0.12.2
PyHamcrest==1.9.0
pytest==5.1.3
pytest-cov==2.7.1
pytest-pythonpath==0.7.3
pytest-sugar==0.9.2
Faker==2.0.2
requests==2.22.0
